+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.14.0] - 2026-03-11
+
 ### Added
 - **restrictDefaultNacl property for VPC construct**
   - Lock down the VPC's default Network ACL to allow VPC CIDR traffic and controlled internet access